refactor(lint): split monolithic test file into per-rule-module test files
